Moving import-fee-periods to the deploy script (for use with generateSolidity)
