+++ publish/src/commands/deploy/index.js
@@ -27,25 +27,26 @@ const {
 	},
 } = require('../../../..');
 
-const performSafetyChecks = require('./perform-safety-checks');
-const getDeployParameterFactory = require('./get-deploy-parameter-factory');
-const systemAndParameterCheck = require('./system-and-parameter-check');
-const deployCore = require('./deploy-core');
-const deploySynths = require('./deploy-synths');
-const deployLoans = require('./deploy-loans');
-const deployDappUtils = require('./deploy-dapp-utils.js');
-const deployBinaryOptions = require('./deploy-binary-options');
-const importAddresses = require('./import-addresses');
-const rebuildResolverCaches = require('./rebuild-resolver-caches');
-const configureLegacySettings = require('./configure-legacy-settings');
-const configureStandalonePriceFeeds = require('./configure-standalone-price-feeds');
-const configureSynths = require('./configure-synths');
 const addSynthsToProtocol = require('./add-synths-to-protocol');
 const configureInverseSynths = require('./configure-inverse-synths');
-const configureSystemSettings = require('./configure-system-settings');
+const configureLegacySettings = require('./configure-legacy-settings');
 const configureLoans = require('./configure-loans');
-const takeDebtSnapshotWhenRequired = require('./take-debt-snapshot-when-required');
+const configureStandalonePriceFeeds = require('./configure-standalone-price-feeds');
+const configureSynths = require('./configure-synths');
+const configureSystemSettings = require('./configure-system-settings');
+const deployBinaryOptions = require('./deploy-binary-options');
+const deployCore = require('./deploy-core');
+const deployDappUtils = require('./deploy-dapp-utils.js');
+const deployLoans = require('./deploy-loans');
+const deploySynths = require('./deploy-synths');
 const generateSolidityOutput = require('./generate-solidity-output');
+const getDeployParameterFactory = require('./get-deploy-parameter-factory');
+const importAddresses = require('./import-addresses');
+const importFeePeriods = require('./import-fee-periods');
+const performSafetyChecks = require('./perform-safety-checks');
+const rebuildResolverCaches = require('./rebuild-resolver-caches');
+const systemAndParameterCheck = require('./system-and-parameter-check');
+const takeDebtSnapshotWhenRequired = require('./take-debt-snapshot-when-required');
 
 const DEFAULTS = {
 	gasPrice: '1',
@@ -227,6 +228,7 @@ const deploy = async ({
 		currentWeekOfInflation,
 		oldExrates,
 		oracleAddress,
+		systemSuspended,
 	} = await systemAndParameterCheck({
 		account,
 		buildPath,
@@ -301,8 +303,10 @@ const deploy = async ({
 		config,
 		deployer,
 		freshDeploy,
+		generateSolidity,
 		network,
 		synths,
+		systemSuspended,
 		yes,
 	});
 
@@ -357,6 +361,17 @@ const deploy = async ({
 		network,
 		runStep,
 		useOvm,
+	});
+
+	await importFeePeriods({
+		deployer,
+		explorerLinkPrefix,
+		freshDeploy,
+		generateSolidity,
+		network,
+		runStep,
+		systemSuspended,
+		yes,
 	});
 
 	await configureStandalonePriceFeeds({
