+++ publish/src/commands/deploy/system-and-parameter-check.js
@@ -241,5 +241,6 @@ module.exports = async ({
 		currentWeekOfInflation,
 		oldExrates,
 		oracleAddress,
+		systemSuspended,
 	};
 };
